Major refactoring of the codebase to improve modularity and readability. This commit includes the following changes:
