+++ app_pages/donorspage_perdonor.py
@@ -1,7 +1,7 @@
 import streamlit as st
-import calculations as ppcalc
+import components.calculations as ppcalc
 import datetime as dt
-import Visualisations as vis
+import components.Visualisations as vis
 import pandas as pd
