+++ app_pages/visits.py
@@ -3,8 +3,8 @@ def visits_body():
     Displays the content of Visit Donations to Political Party page.
     """
     import streamlit as st
-    import calculations as ppcalc
-    import Visualisations as vis
+    import components.calculations as ppcalc
+    import components.Visualisations as vis
     import datetime as dt
 
     # Load dataset from session state
